Progress on J2DWindow

+++ include/JSystem/JUtility/JUTPalette.hpp
@@ -26,7 +26,7 @@ public:
 
 	u32 getTlutName() const { return mTlutName; }
 	u8 getFormat() const { return mFormat; }
-	u8 getTransparency() const { return mTransparency; }
+	u32 getTransparency() const { return mTransparency; }
 	u16 getNumColors() const { return mNumColors; }
 	ResTLUT* getColorTable() const { return mColorTable; }
